API_and_DB/requirements.txt
flask
flask-cors
gunicorn
huggingface-hub
openai-whisper
numpy
torch
torchaudio
pymongo
dnspython
Pillow
requests
dotenv
python-dotenv